Added resolveDate() and formArrBuilder()
My Sheet has a column with a due date in a number and trailing letter representing an amount of time between the created date and the unprovided due date. (7d, 2w, 13d, etc...). I've added a function to add Excel logic to accept an array built from formArrBuilder() into the last column (offset) to add the requisite number of days to the created date.  I then edited the formulas from calendarDay() and lastCalendarDay() to accommodate for the resolved date.